requirements.txt
# requirements.txt
# Core ML/NLP
sentence-transformers==2.2.2
scikit-learn==1.3.2
hdbscan==0.8.33
umap-learn==0.5.5

# LLM & Agents
langgraph==0.2.0
langchain==0.2.0
langchain-anthropic==0.1.15
langchain-core==0.2.0

# Visualization & UI
streamlit==1.28.0
plotly==5.18.0

# Data
pandas==2.1.3
numpy==1.24.3

# Optional: For real web search (uncomment if using)
# tavily-python==0.3.0
# google-search-results==2.4.2